backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.11.0",
    "boto3>=1.42.24",
    "django==5.2.9",
    "django-cors-headers==4.5.0",
    "django-filter>=25.2",
    "django-rest-framework>=0.1.0",
    "django-storages>=1.14.6",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.4.0",
    "drf-spectacular==0.29.0",
    "minio>=7.2.20",
    "python-dotenv>=1.2.1",
    "rest-framework-simplejwt>=0.0.2",
    "sqlparse==0.5.4",
]
